requirements.txt
torch
torchvision
onnx
onnxruntime
onnxruntime-tools
numpy
pandas
matplotlib
Pillow
opencv-python
facenet-pytorch
gradio
gdown